+++ CHANGELOG.md
@@ -1,5 +1,49 @@
 # Changelog
 
+## 1.5.1 - 2016-10-28
+
+- Only set attribute if field is bound, see #567
+
+## 1.5.0 - 2016-10-27
+
+- Make Pikaday accessibile, see #522
+- Add possibility to reset the max and min date to the date picker
+- Make Pikaday XHTML compatible, see #492
+- Remove unnecessary setTimeout call, fixes #496
+- Make disabled buttons not clickable in IE 10 and below, fixes #288
+- Show days in next and previous months
+
+## 1.4.0 - 2015-10-30
+
+- Fix touch device support, see #374
+- Add possibility to reset the max and min date to the date picker
+
+## 1.3.3 - 2015-02-17
+
+- Add theme class support, see #260
+- Fix on iOS, see #101, #121
+- Add date range example, see #337
+- Fix dates used in examples, see #329
+
+## 1.3.2 - 2015-02-17
+
+- Add options to disable days, see #171
+- Add SCSS file, see #130
+
+## 1.3.1 - 2015-02-04
+
+- Fix for bug introduced in #166; fixes #224, #230
+
+## 1.3.0 - 2015-01-09
+
+- Fix #79 moving through months can cause calendar to "bounce around"
+- Add configurable number of shown months
+- #51 Add unit tests with ci: [testling](https://ci.testling.com/rikkert/pikaday)
+- Fix #94 blur input field after date selection
+- Add configuable container to render calendar into
+- Add option to show week numbers, see #147
+- Positioning is now optional, see #192
+
 ## 1.2.0 - 2014-01-28
 
 - Fix #8 Automatically adjust datepicker position